Saturday night commit; added basic title screen and more fixes to combat system

+++ Assets/Scripts/CombatSimple/Data/ActionsBehaviourEntry.cs
@@ -3,7 +3,7 @@
     [System.Serializable]
     public class ActionsBehaviourEntry
     {
-        public float actionPointsGain;
+        public int actionPointsGain;
         public Dialogue dialogue;
     }
 }
